no need for uint32 on createOffers

+++ contracts/modules/CollectionOffers/V1/CollectionOffersV1.sol
@@ -128,12 +128,12 @@ contract CollectionOffersV1 is
     /// @notice Creates an offer for any NFT in a collection
     /// @param _tokenContract The ERC-721 collection address
     /// @return The ID of the created offer
-    function createOffer(address _tokenContract) external payable nonReentrant returns (uint32) {
+    function createOffer(address _tokenContract) external payable nonReentrant returns (uint256) {
         // Ensure offer is valid and take custody
         _handleIncomingTransfer(msg.value, ETH);
 
         // Add to collection's offer book
-        uint32 offerId = _addOffer(_tokenContract, msg.value, msg.sender);
+        uint256 offerId = _addOffer(_tokenContract, msg.value, msg.sender);
 
         emit CollectionOfferCreated(_tokenContract, offerId, msg.sender, msg.value);
